moved workspaces pkg out of core
